moving utils away from prettier-comments folder since now is slim and understandable enough to be maintained.
